+++ release_notes.md
@@ -6,6 +6,7 @@
 * renamed DotnetCompile to DotnetBuild
 * updated Dotnet* helpers to match current rtw cli options (now based on msbuild)
 * multiple renames to rebrand some DotnetCli helpers to DotnetSdk
+* make CommonOptions function
 
 #### 1.1.1
 * add preview2 tooling .NET Core 1.0.1 shortcut (Preview2Tooling101Options)